Compile but errors
Server and code are compiling, however there are still some errors in Json Parsing.
